requirements.txt
streamlit
openai
gtts
audio-recorder-streamlit
mysql-connector-python
python-dotenv